Parallelize using only Thread class. Calculate the sum of squares of the given ranges.
